Create program for snakes and ladders
Description:

Snakes and Ladder is the most common board game played. The rules of the game are as follows:

1. The first person to reach 100 wins.
2. Each player gets only one chance in a single traversing.
3. Snakes decrease your points while the ladder increases them.
So, as the rules are quite easy to understand we can easily code them using C language to create a Snake and Ladder Application. The functionality of the code will be as follows:

1. Two players can enter a single game.
2. Random values can be attained using dice to increase or decrease the value.
3. Points will be maintained using variables.
4. The game will end after any player attains 100 points.